Prometheus/docker/applications/python/frontend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
grpcio==1.70.0
h11==0.14.0
httpx==0.28.1
idna==3.10
opentelemetry-distro==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-logging==0.51b0
opentelemetry-exporter-otlp==1.30.0
prometheus_client==0.21.1
psutil==6.1.1
py-grpc-prometheus==0.8.0
pydantic==2.10.6
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
